RAC-351 fix : 중복코드 제거
